requirements.txt
aiogram==3.15.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
apscheduler==3.10.4
python-dotenv==1.0.1
